LayoutTests/imported/w3c/web-platform-tests/html/anonymous-iframe/cookie.tentative.https.window-expected.txt


PASS Setup
FAIL Credentialless same-origin iframe is loaded without credentials assert_equals: expected (undefined) undefined but got (string) "same_origin"
FAIL Credentialless cross-origin iframe is loaded without credentials assert_equals: expected (undefined) undefined but got (string) "cross_origin"
FAIL same_origin credentialless iframe can't send same_origin credentials assert_equals: expected (undefined) undefined but got (string) "same_origin"
FAIL same_origin credentialless iframe can't send cross_origin credentials assert_equals: expected (undefined) undefined but got (string) "cross_origin"
FAIL cross_origin credentialless iframe can't send cross_origin credentials assert_equals: expected (undefined) undefined but got (string) "cross_origin"
FAIL cross_origin credentialless iframe can't send same_origin credentials assert_equals: expected (undefined) undefined but got (string) "same_origin"
FAIL same_origin credentialless iframe can't send same_origin credentials on child iframe assert_equals: expected (undefined) undefined but got (string) "same_origin"
FAIL same_origin credentialless iframe can't send cross_origin credentials on child iframe assert_equals: expected (undefined) undefined but got (string) "cross_origin"
FAIL cross_origin credentialless iframe can't send cross_origin credentials on child iframe assert_equals: expected (undefined) undefined but got (string) "cross_origin"
FAIL cross_origin credentialless iframe can't send same_origin credentials on child iframe assert_equals: expected (undefined) undefined but got (string) "same_origin"
